+++ changelog.md
@@ -5,6 +5,7 @@
 
 ##### New Features
 * Add ability to connect to an energy workbench server without authentication.
+* `EasClient.upload_study` now returns a HTTP response.
 
 ##### Enhancements
 * None.